Apply referrer policy to links only #7
Meta referrer affects more than just links, and #5 and #7 are examples
of why this is not good. Some Google API endpoints check the referrer
header and reject the request if the referrer does not match.

With this commit, referrer headers on outgoing links are cleaned via an
explicit referrer policy on the link itself instead of a global meta
tag. I initially used this meta tag because Gmail opened links via
iframes. But with the patch for #2 and the extra hooks on
`dispatchEvent` and `click`, assigning the referrer policy via links
ought to work in most if not all cases.

referrerPolicy for anchors is supported by default in Firefox 50+ (which
was officially released last week; before that the feature can be
enabled via the `network.http.enablePerElementReferrer` pref) and
Chrome 51+.

Tested as follows:
- Search for "example" in Google, right-click link, open in new tab via
  context menu. Open devtools and confirm that `document.referrer` only
  exposes the origin (not the search terms).

- Open devtools at Google, try the following snippets and confirm that
  the result is always a link with the referrerpolicy attribute:

  * `var a = document.createElement('a');a.href='';a`
  * `var a = document.createElement('a');a.click();a`
  * `var a = document.createElement('a');a.dispatchEvent(new CustomEvent(''));a`