Account for CSP when running scripts in page
Google may have a stricter CSP that blocks inline scripts, unless a
nonce is specified. To avoid triggering CSP violation reports, we
assume that the page has a strict CSP in Firefox, and await the presence
of a CSP nonce, and use it to execute scripts anyway.

In Chrome, the behavior hasn't changed significantly.