Sanitize meta refresh popups (Docs) #41
When a link is clicked in Google Docs, an inline box is opened. After
clicking on the displayed link in that box, the link target is opened
in a new window/tab.

Google Docs implements this by intercepting a "click" and using
`window.open` to open a blank popup and using meta-refresh to navigate
to an intermediate page (`/url?q=...`) instead of the final destination:
https://github.com/Rob--W/dont-track-me-google/issues/41#issuecomment-780835972

This is resolved by intercepting the call and sanitizing the URL.