Unbreak Presenter View in Google Slides #49
This improves the logic originally introduced for #41.

The noreferrer flag was already skipped for missing URLs (empty
strings), but not for "about:blank" yet. The two are equivalent,
so they should be treated as such.

This fixes a bug on Google Slides, where the Presenter View fails to
render speaker notes because Docs was unable to get a window reference
to the new popup due to the "noopener" flag implied by "noreferrer".