Overwrite anchor prototype without triggering setters #52
Assignment using `proto.dispatchEvent = ...` triggers can trigger
setters, e.g. on console.cloud.google.com (#52). In this case, it was
observed that the page redirected the assignment to the original
interface on which the method was defined, instead of
HTMLAnchorElement.prototype.

This patch replaces the method specifically on HTMLAnchorElement to
avoid the mentioned issue, which avoids performance degradation.