Replace extension preference retrieval mechanism
In the investigation of #52 I discovered that console.cloud.google.com
replaced many methods of EventTarget.prototype (and Element.prototype).

Because content script is not guaranteed to run before the page, that
caused infinite recursion when `dispatchEvent` in `updateReferrerPolicy`
was redirected back to `proto.dispatchEvent` as explained at
https://github.com/Rob--W/dont-track-me-google/issues/52#issuecomment-1596207655

This commit works around the specific issue by using a specific part of
the HTMLScriptElement interface to communicate the referrer policy to
the extension.

All uses of CustomEvent/dispatchEvent have been replaced too; the
content script now switched from a pull-config to a push-config
approach. This approach is also more optimal because the config rarely
changes.