Move control of cleanLinksWhenJsIsDisabled to contentscript.js
The execution of cleanLinksWhenJsIsDisabled in contentscript.js
should be independent of main_world_script.js. It was there
because the logic was originally called when jsEnabled is false,
but later it was also unconditionally enabled for Google Docs.

Now this logic fits better in contentscript.js, so we've
introduced the potentiallyAsyncInit() function that calls
cleanLinksWhenJsIsDisabled() when needed.

As a bonus, it is called after we've read the storage to ensure that the
logic runs with the latest known user preferences.